fix running break on macosx

+++ smartcut/cut_video.py
@@ -481,11 +481,11 @@ class VideoCutter:
                 self.transform_graph.vpush(frame)
                 frame = self.transform_graph.vpull()
 
-            frame.pts -= s.start_time / in_tb
+            frame.pts -= int(s.start_time / in_tb)
 
-            frame.pts = frame.pts * in_tb / out_tb
+            frame.pts = int(frame.pts * in_tb / out_tb)
             frame.time_base = out_tb
-            frame.pts += self.segment_start_in_output / out_tb
+            frame.pts += int(self.segment_start_in_output / out_tb)
 
             if frame.pts <= self.enc_last_pts:
                 frame.pts = self.enc_last_pts + 1
@@ -511,12 +511,12 @@ class VideoCutter:
         for packet in itertools.chain([packet], demux_iter):
             last_packet = packet.dts == s.gop_end_dts
             packet.pts -= segment_start_pts
-            packet.pts = packet.pts * self.in_stream.time_base / self.out_time_base
-            packet.pts += self.segment_start_in_output / self.out_time_base
+            packet.pts = int(packet.pts * self.in_stream.time_base / self.out_time_base)
+            packet.pts += int(self.segment_start_in_output / self.out_time_base)
             if packet.dts is not None:
                 packet.dts -= segment_start_pts
-                packet.dts = packet.dts * self.in_stream.time_base / self.out_time_base
-                packet.dts += self.segment_start_in_output / self.out_time_base
+                packet.dts = int(packet.dts * self.in_stream.time_base / self.out_time_base)
+                packet.dts += int(self.segment_start_in_output / self.out_time_base)
 
             result_packets.extend(self.remux_bitstream_filter.filter(packet))
